+++ src/main/java/dds/monedero/model/Extraccion.java
@@ -5,7 +5,7 @@ import java.time.LocalDate;
 
 public class Extraccion extends Movimiento {
 
-  public Extraccion(LocalDate fecha, BigDecimal monto) {
+  public Extraccion(LocalDate fecha, double monto) {
     super(fecha, monto);
   }
